Get RandomSource facility working under Hugs.

+++ src/Language/Robin/Facilities/RandomSource.hs
@@ -13,6 +13,6 @@ init = return Facility{ handler=handleEvent, waiter=nullWaiter }
 
 handleEvent :: FacilityHandler
 handleEvent (List [Symbol "obtain-random-u16", payload]) = do
-    v <- randomRIO (0, 65535) :: IO Int32
-    return $ [List [Symbol "random-u16", Number v]]
+    v <- randomRIO (0, 65535) :: IO Integer
+    return $ [List [Symbol "random-u16", Number (fromIntegral v)]]
 handleEvent _ = return []
